requirements.txt
google-generativeai>=0.4.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
click>=8.0.0
rich>=10.0.0
tabulate>=0.9.0
sqlparse>=0.4.0
streamlit>=1.28.0
pandas>=1.5.0
asyncio-compat>=0.1.0 